+++ src/conn.rs
@@ -1,18 +1,7 @@
 use mio::Poll;
-use mio::PollOpt;
-use mio::Ready;
 use mio::Token;
-use mio::unix::EventedFd;
-use net2::TcpBuilder;
-use net2::TcpStreamExt;
 use std::collections::HashSet;
-use std::io::Read;
-use std::io::Write;
-use std::io;
-use std::net::SocketAddr;
-use std::net::TcpStream;
-use std::net::ToSocketAddrs;
-use std::os::unix::io::{AsRawFd, RawFd};
+use std::result;
 use std::str;
 
 use config;
@@ -21,10 +10,12 @@ use logger::Logger;
 use utils;
 use wire::{Cmd, Msg, Pfx};
 use wire;
+use stream::{Stream, StreamErr};
 
 pub struct Conn<'poll> {
     serv_addr: String,
     serv_port: u16,
+    tls: bool,
     hostname: String,
     realname: String,
     nicks: Vec<String>,
@@ -50,33 +41,18 @@ pub struct Conn<'poll> {
     /// don't parse USERHOST responses to set this field.
     usermask: Option<String>,
 
-    /// The TCP connection to the server.
-    stream: TcpStream,
-
     poll: &'poll Poll,
 
-    status: ConnStatus,
+    status: ConnStatus<'poll>,
 
     /// Incoming message buffer
     in_buf: Vec<u8>,
-
-    /// Outgoing message buffer. Make sure to register the socket for rw events
-    /// after writing to this buffer.
-    out_buf: Vec<u8>,
 }
 
-fn deregister_fd(poll: &Poll, fd: RawFd) {
-    // deregistering multiple times is fine .. I think
-    // plus we call this in drop() so we shoudn't panic
-    let _ = poll.deregister(&EventedFd(&fd));
-}
+pub type ConnErr = StreamErr;
 
-impl<'poll> Drop for Conn<'poll> {
-    fn drop(&mut self) {
-        deregister_fd(self.poll, self.get_raw_fd());
-    }
-}
-
+/// How many ticks to wait before assuming disconnect in introduce state.
+const INTRO_TICKS: u8 = 30;
 /// How many ticks to wait before sending a ping to the server.
 const PING_TICKS: u8 = 60;
 /// How many ticks to wait after sending a ping to the server to consider a
@@ -85,23 +61,64 @@ const PONG_TICKS: u8 = 60;
 /// How many ticks to wait after a disconnect or a socket error.
 pub const RECONNECT_TICKS: u8 = 30;
 
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-enum ConnStatus {
+enum ConnStatus<'poll> {
     /// Need to introduce self
-    Introduce,
+    Introduce {
+        ticks_passed: u8,
+        stream: Stream<'poll>,
+    },
     PingPong {
         /// Ticks passed since last time we've heard from the server. Reset on
         /// each message. After `PING_TICKS` ticks we send a PING message and
         /// move to `WaitPong` state.
         ticks_passed: u8,
+        stream: Stream<'poll>,
     },
     WaitPong {
         /// Ticks passed since we sent a PING to the server. After a message
         /// move to `PingPong` state. On timeout we reset the connection.
         ticks_passed: u8,
+        stream: Stream<'poll>,
     },
     Disconnected { ticks_passed: u8 },
 }
+
+macro_rules! update_status {
+    ($self:ident, $v:ident, $code:expr) => {{
+        // temporarily putting `Disconnected` to `self.status`
+        let $v = ::std::mem::replace(&mut $self.status, ConnStatus::Disconnected { ticks_passed: 0 });
+        let new_status = $code;
+        $self.status = new_status;
+    }}
+}
+
+impl<'poll> ConnStatus<'poll> {
+    fn get_stream(&self) -> Option<&Stream<'poll>> {
+        use self::ConnStatus::*;
+        match *self {
+            Introduce { ref stream, .. } |
+            PingPong { ref stream, .. } |
+            WaitPong { ref stream, .. } =>
+                Some(stream),
+            Disconnected { .. } =>
+                None,
+        }
+    }
+
+    fn get_stream_mut(&mut self) -> Option<&mut Stream<'poll>> {
+        use self::ConnStatus::*;
+        match *self {
+            Introduce { ref mut stream, .. } |
+            PingPong { ref mut stream, .. } |
+            WaitPong { ref mut stream, .. } =>
+                Some(stream),
+            Disconnected { .. } =>
+                None,
+        }
+    }
+}
+
+pub type Result<T> = result::Result<T, StreamErr>;
 
 #[derive(Debug)]
 pub enum ConnEv {
@@ -112,79 +129,21 @@ pub enum ConnEv {
     /// Hack to return the main loop that the Conn wants reconnect()
     WantReconnect,
     /// Network error happened
-    Err(io::Error),
+    Err(StreamErr),
     /// An incoming message
     Msg(Msg),
     /// Nick changed
     NickChange(String),
 }
 
-fn init_stream(serv_addr: &str, serv_port: u16) -> TcpStream {
-    // FIXME: This socket will return an error and make the Conn enter into
-    // "reconnnect" loop. There should be a more direct way of doing this.
-    fn mk_useless_stream() -> TcpStream {
-        let stream = TcpBuilder::new_v4().unwrap().to_tcp_stream().unwrap();
-        stream.set_nonblocking(true).unwrap();
-        stream
-    }
-
-    // FIXME: This will block the thread. See issue #3.
-    // FIXME: This part is really horrible. No way to report errors. The `Conn`
-    // will just try to reconnect in case of an error.
-    match (serv_addr, serv_port).to_socket_addrs() {
-        Err(_) =>
-            mk_useless_stream(),
-        Ok(mut addr_iter) => {
-            match addr_iter.next() {
-                None =>
-                    mk_useless_stream(),
-                Some(SocketAddr::V4(addr)) => {
-                    let stream = TcpBuilder::new_v4().unwrap().to_tcp_stream().unwrap();
-                    stream.set_nonblocking(true).unwrap();
-                    // This will fail with EINPROGRESS
-                    let _ = stream.connect(SocketAddr::V4(addr));
-                    stream
-                }
-                Some(SocketAddr::V6(addr)) => {
-                    let stream = TcpBuilder::new_v6().unwrap().to_tcp_stream().unwrap();
-                    stream.set_nonblocking(true).unwrap();
-                    // This will fail with EINPROGRESS
-                    let _ = stream.connect(SocketAddr::V6(addr));
-                    stream
-                }
-            }
-        }
-    }
-}
-
-fn reregister_for_rw(poll: &Poll, fd: RawFd) {
-    // fails when not already registered, ignore result
-    // (e.g. between a disconnect and reconnect)
-    let _ = poll.reregister(
-        &EventedFd(&fd),
-        Token(fd as usize),
-        Ready::readable() | Ready::writable(),
-        PollOpt::level(),
-    );
-}
-
-fn reregister_for_r(poll: &Poll, fd: RawFd) {
-    // fails when not already registered, ignore result
-    // (e.g. between a disconnect and reconnect)
-    let _ = poll.reregister(
-        &EventedFd(&fd),
-        Token(fd as usize),
-        Ready::readable(),
-        PollOpt::level(),
-    );
-}
-
 impl<'poll> Conn<'poll> {
-    pub fn from_server(server: config::Server, poll: &'poll Poll) -> Conn {
-        let stream = init_stream(&server.addr, server.port);
-        let ret = Conn {
+    pub fn from_server(server: config::Server, poll: &'poll Poll) -> Result<Conn<'poll>> {
+        let stream =
+            Stream::new(poll, &server.addr, server.port, server.tls).map_err(StreamErr::from)?;
+        Ok(Conn {
             serv_addr: server.addr,
             serv_port: server.port,
+            tls: server.tls,
             hostname: server.hostname,
             realname: server.realname,
             nicks: server.nicks,
@@ -193,88 +152,71 @@ impl<'poll> Conn<'poll> {
             away_status: None,
             servername: None,
             usermask: None,
-            stream: stream,
             poll: poll,
-            status: ConnStatus::Introduce,
+            status: ConnStatus::Introduce {
+                ticks_passed: 0,
+                stream: stream,
+            },
             in_buf: vec![],
-            out_buf: vec![],
-        };
-        ret.register_for_r();
-        ret
+        })
     }
 
     /// Clone an existing connection, but update the server address.
-    pub fn from_conn(conn: &Conn<'poll>, new_serv_addr: &str, new_serv_port: u16) -> Conn<'poll> {
-        let stream = init_stream(new_serv_addr, new_serv_port);
-        let ret = Conn {
+    pub fn from_conn(
+        conn: Conn<'poll>,
+        new_serv_addr: &str,
+        new_serv_port: u16,
+    ) -> Result<Conn<'poll>> {
+        Ok(Conn {
             serv_addr: new_serv_addr.to_owned(),
             serv_port: new_serv_port,
-            hostname: conn.hostname.clone(),
-            realname: conn.realname.clone(),
+            tls: conn.tls,
+            hostname: conn.hostname,
+            realname: conn.realname,
             nicks: conn.nicks.clone(),
             current_nick_idx: 0,
             auto_join: HashSet::new(),
             away_status: None,
             servername: None,
             usermask: None,
-            stream: stream,
             poll: conn.poll,
-            status: ConnStatus::Introduce,
+            status: ConnStatus::Introduce {
+                ticks_passed: 0,
+                stream: Stream::new(conn.poll, new_serv_addr, new_serv_port, conn.tls)
+                    .map_err(StreamErr::from)?,
+            },
             in_buf: vec![],
-            out_buf: vec![],
-        };
-        ret.register_for_r();
-        ret
+        })
     }
 
-    /// Register self to the Poll for read events.
-    fn register_for_r(&self) {
-        let fd = self.get_raw_fd();
-        self.poll
-            .register(
-                &EventedFd(&fd),
-                Token(fd as usize),
-                Ready::readable(),
-                PollOpt::level(),
-            )
-            .unwrap();
-    }
+    pub fn reconnect(&mut self, new_serv: Option<(&str, u16)>) -> Result<()> {
+        // drop existing connection first
+        let old_stream = ::std::mem::replace(
+            &mut self.status,
+            ConnStatus::Disconnected { ticks_passed: 0 },
+        );
+        drop(old_stream);
 
-    /// Re-register self to the Poll for read events.
-    fn reregister_for_r(&self) {
-        reregister_for_r(self.poll, self.get_raw_fd());
-    }
-
-    /// Re-register self to the Poll for read and write events.
-    fn reregister_for_rw(&self) {
-        reregister_for_rw(self.poll, self.get_raw_fd());
-    }
-
-    /// De-register self. Do this after a connection error.
-    fn deregister(&self) {
-        deregister_fd(self.poll, self.get_raw_fd());
-    }
-
-    pub fn reconnect(&mut self, new_serv: Option<(&str, u16)>) {
-        self.deregister();
         if let Some((new_name, new_port)) = new_serv {
             self.serv_addr = new_name.to_owned();
             self.serv_port = new_port;
         }
-        self.stream = init_stream(&self.serv_addr, self.serv_port);
-        self.status = ConnStatus::Introduce;
-        self.current_nick_idx = 0;
-        self.out_buf.clear();
-        self.register_for_r();
+        match Stream::new(self.poll, &self.serv_addr, self.serv_port, self.tls) {
+            Err(err) =>
+                Err(StreamErr::from(err)),
+            Ok(stream) => {
+                self.status = ConnStatus::Introduce {
+                    ticks_passed: 0,
+                    stream: stream,
+                };
+                self.current_nick_idx = 0;
+                Ok(())
+            }
+        }
     }
 
-    /// Get the RawFd, to be used with select() or other I/O multiplexer.
-    fn get_raw_fd(&self) -> RawFd {
-        self.stream.as_raw_fd()
-    }
-
-    pub fn get_conn_tok(&self) -> Token {
-        Token(self.get_raw_fd() as usize)
+    pub fn get_conn_tok(&self) -> Option<Token> {
+        self.status.get_stream().map(|s| s.get_tok())
     }
 
     pub fn get_serv_name(&self) -> &str {
@@ -295,10 +237,6 @@ impl<'poll> Conn<'poll> {
         self.send_nick();
     }
 
-    fn reset_nick(&mut self) {
-        self.current_nick_idx = 0;
-    }
-
     fn next_nick(&mut self) {
         if self.current_nick_idx + 1 == self.nicks.len() {
             let mut new_nick = self.nicks.last().unwrap().to_string();
@@ -310,86 +248,123 @@ impl<'poll> Conn<'poll> {
 }
 
 impl<'poll> Conn<'poll> {
+    pub fn enter_disconnect_state(&mut self) {
+        self.status = ConnStatus::Disconnected { ticks_passed: 0 };
+    }
+
     ////////////////////////////////////////////////////////////////////////////
     // Tick handling
 
     pub fn tick(&mut self, evs: &mut Vec<ConnEv>, mut debug_out: LogFile) {
-        match self.status {
-            ConnStatus::Introduce =>
-                {}
-            ConnStatus::PingPong { ticks_passed } =>
-                if ticks_passed + 1 == PING_TICKS {
-                    match self.servername {
-                        None => {
-                            debug_out.write_line(format_args!(
-                                "{}: Can't send PING, servername unknown",
-                                self.serv_addr
-                            ));
-                        }
-                        Some(ref host_) => {
-                            debug_out.write_line(format_args!(
-                                "{}: Ping timeout, sending PING",
-                                self.serv_addr
-                            ));
-                            wire::ping(&mut self.out_buf, host_).unwrap();
-                            reregister_for_rw(self.poll, self.get_raw_fd());
+        update_status!(
+            self,
+            status,
+            match status {
+                ConnStatus::Introduce {
+                    stream,
+                    ticks_passed,
+                } => {
+                    let ticks = ticks_passed + 1;
+                    if ticks == INTRO_TICKS {
+                        evs.push(ConnEv::Disconnected);
+                        ConnStatus::Disconnected { ticks_passed: 0 }
+                    } else {
+                        ConnStatus::Introduce {
+                            stream,
+                            ticks_passed: ticks,
                         }
                     }
-                    self.status = ConnStatus::WaitPong { ticks_passed: 0 };
-                } else {
-                    self.status = ConnStatus::PingPong {
-                        ticks_passed: ticks_passed + 1,
-                    };
-                },
-            ConnStatus::WaitPong { ticks_passed } =>
-                if ticks_passed + 1 == PONG_TICKS {
-                    evs.push(ConnEv::Disconnected);
-                    self.enter_disconnect_state();
-                } else {
-                    self.status = ConnStatus::WaitPong {
-                        ticks_passed: ticks_passed + 1,
-                    };
-                },
-            ConnStatus::Disconnected { ticks_passed } => {
-                if ticks_passed + 1 == RECONNECT_TICKS {
-                    // *sigh* it's slightly annoying that we can't reconnect here, we need to
-                    // update the event loop
-                    evs.push(ConnEv::WantReconnect);
-                    self.reset_nick();
                 }
-                self.status = ConnStatus::Disconnected {
-                    ticks_passed: ticks_passed + 1,
-                };
+                ConnStatus::PingPong {
+                    mut stream,
+                    ticks_passed,
+                } => {
+                    let ticks = ticks_passed + 1;
+                    if ticks == PING_TICKS {
+                        match self.servername {
+                            None => {
+                                debug_out.write_line(format_args!(
+                                    "{}: Can't send PING, servername unknown",
+                                    self.serv_addr
+                                ));
+                            }
+                            Some(ref host_) => {
+                                debug_out.write_line(format_args!(
+                                    "{}: Ping timeout, sending PING",
+                                    self.serv_addr
+                                ));
+                                wire::ping(&mut stream, host_).unwrap();
+                            }
+                        }
+                        ConnStatus::WaitPong {
+                            stream,
+                            ticks_passed: 0,
+                        }
+                    } else {
+                        ConnStatus::PingPong {
+                            stream,
+                            ticks_passed: ticks,
+                        }
+                    }
+                }
+                ConnStatus::WaitPong {
+                    stream,
+                    ticks_passed,
+                } => {
+                    let ticks = ticks_passed + 1;
+                    if ticks == PONG_TICKS {
+                        evs.push(ConnEv::Disconnected);
+                        ConnStatus::Disconnected { ticks_passed: 0 }
+                    } else {
+                        ConnStatus::WaitPong {
+                            stream,
+                            ticks_passed: ticks,
+                        }
+                    }
+                }
+                ConnStatus::Disconnected { ticks_passed } => {
+                    let ticks = ticks_passed + 1;
+                    if ticks_passed + 1 == RECONNECT_TICKS {
+                        // *sigh* it's slightly annoying that we can't reconnect here, we need to
+                        // update the event loop
+                        evs.push(ConnEv::WantReconnect);
+                        self.current_nick_idx = 0;
+                    }
+                    ConnStatus::Disconnected {
+                        ticks_passed: ticks,
+                    }
+                }
             }
-        }
-    }
-
-    pub fn enter_disconnect_state(&mut self) {
-        self.status = ConnStatus::Disconnected { ticks_passed: 0 };
-        self.deregister();
+        );
     }
 
     fn reset_ticks(&mut self) {
-        match self.status {
-            ConnStatus::Introduce =>
-                {}
-            _ => {
-                self.status = ConnStatus::PingPong { ticks_passed: 0 };
+        update_status!(
+            self,
+            status,
+            match status {
+                ConnStatus::Introduce { stream, .. } =>
+                    ConnStatus::Introduce { ticks_passed: 0, stream },
+                ConnStatus::PingPong { stream, .. } =>
+                    ConnStatus::PingPong { ticks_passed: 0, stream },
+                ConnStatus::WaitPong { stream, .. } =>
+                    // no bug: we heard something from the server, whether it was a pong or not
+                    // doesn't matter that much, connectivity is fine.
+                    ConnStatus::PingPong { ticks_passed: 0, stream },
+                ConnStatus::Disconnected { .. } =>
+                    status,
             }
-        }
+        );
     }
 
     ////////////////////////////////////////////////////////////////////////////
     // Sending messages
 
-    fn introduce(&mut self) {
-        wire::user(&mut self.out_buf, &self.hostname, &self.realname).unwrap();
-        self.send_nick();
-    }
-
     fn send_nick(&mut self) {
-        wire::nick(&mut self.out_buf, &self.nicks[self.current_nick_idx]).unwrap();
-        self.reregister_for_rw();
+        let nick = &self.nicks[self.current_nick_idx];
+        self.status.get_stream_mut().map(|stream| {
+            wire::nick(stream, nick).unwrap();
+        });
     }
 
     /// `extra_len`: Size (in bytes) for a prefix/suffix etc. that'll be added to each line.
@@ -423,47 +398,51 @@ impl<'poll> Conn<'poll> {
     // to fit into 512 bytes. Need to make sure `split_privmsg` is called before
     // this.
     pub fn privmsg(&mut self, target: &str, msg: &str) {
-        wire::privmsg(&mut self.out_buf, target, msg).unwrap();
-        self.reregister_for_rw();
+        self.status.get_stream_mut().map(|stream| {
+            wire::privmsg(stream, target, msg).unwrap();
+        });
     }
 
     pub fn ctcp_action(&mut self, target: &str, msg: &str) {
-        wire::ctcp_action(&mut self.out_buf, target, msg).unwrap();
-        self.reregister_for_rw();
+        self.status.get_stream_mut().map(|stream| {
+            wire::ctcp_action(stream, target, msg).unwrap();
+        });
     }
 
     pub fn join(&mut self, chan: &str) {
-        wire::join(&mut self.out_buf, chan).unwrap();
+        self.status.get_stream_mut().map(|stream| {
+            wire::join(stream, chan).unwrap();
+        });
         // the channel will be added to auto-join list on successful join (i.e.
         // after RPL_TOPIC)
-        self.reregister_for_rw();
     }
 
     pub fn part(&mut self, chan: &str) {
-        wire::part(&mut self.out_buf, chan).unwrap();
-        self.reregister_for_rw();
+        self.status.get_stream_mut().map(|stream| {
+            wire::part(stream, chan).unwrap();
+        });
         self.auto_join.remove(chan);
     }
 
     pub fn away(&mut self, msg: Option<&str>) {
         self.away_status = msg.map(|s| s.to_string());
-        wire::away(&mut self.out_buf, msg).unwrap();
-        self.reregister_for_rw();
+        self.status.get_stream_mut().map(|stream| {
+            wire::away(stream, msg).unwrap();
+        });
     }
 
     ////////////////////////////////////////////////////////////////////////////
     // Sending messages
 
-    pub fn send(&mut self, evs: &mut Vec<ConnEv>) {
-        match self.stream.write(&self.out_buf) {
-            Err(err) => {
-                evs.push(ConnEv::Err(err));
-            }
-            Ok(bytes_sent) => {
-                self.out_buf.drain(0..bytes_sent);
-                if self.out_buf.is_empty() {
-                    self.reregister_for_r();
-                }
+    pub fn write_ready(&mut self, evs: &mut Vec<ConnEv>) {
+        if let Some(stream) = self.status.get_stream_mut() {
+            match stream.write_ready() {
+                Err(err) =>
+                    if !err.is_would_block() {
+                        evs.push(ConnEv::Err(StreamErr::from(err)));
+                    },
+                Ok(()) =>
+                    {}
             }
         }
     }
@@ -471,23 +450,32 @@ impl<'poll> Conn<'poll> {
     ////////////////////////////////////////////////////////////////////////////
     // Receiving messages
 
-    pub fn recv(&mut self, evs: &mut Vec<ConnEv>, logger: &mut Logger) {
+    pub fn read_ready(&mut self, evs: &mut Vec<ConnEv>, logger: &mut Logger) {
         let mut read_buf: [u8; 512] = [0; 512];
 
-        // Handle disconnects
-        match self.stream.read(&mut read_buf) {
-            Err(err) => {
-                evs.push(ConnEv::Err(err));
+        // borrowchk workaround
+        let read_ret = {
+            match self.status.get_stream_mut() {
+                Some(stream) =>
+                    match stream.read_ready(&mut read_buf) {
+                        Err(err) => {
+                            if !err.is_would_block() {
+                                evs.push(ConnEv::Err(StreamErr::from(err)));
+                            }
+                            None
+                        }
+                        Ok(bytes_read) =>
+                            Some(bytes_read),
+                    },
+                None =>
+                    None,
             }
-            Ok(bytes_read) => {
-                self.reset_ticks();
-                self.in_buf.extend(&read_buf[0..bytes_read]);
-                self.handle_msgs(evs, logger);
-                if bytes_read == 0 {
-                    evs.push(ConnEv::Disconnected);
-                    self.enter_disconnect_state();
-                }
-            }
+        };
+
+        if let Some(bytes_read) = read_ret {
+            self.reset_ticks();
+            self.in_buf.extend(&read_buf[0..bytes_read]);
+            self.handle_msgs(evs, logger);
         }
     }
 
@@ -506,15 +494,28 @@ impl<'poll> Conn<'poll> {
             ..
         } = msg
         {
-            wire::pong(&mut self.out_buf, server).unwrap();
-            self.reregister_for_rw();
+            self.status.get_stream_mut().map(|stream| {
+                wire::pong(stream, server).unwrap();
+            });
         }
 
-        if let ConnStatus::Introduce = self.status {
-            self.introduce();
-            self.status = ConnStatus::PingPong { ticks_passed: 0 };
-            evs.push(ConnEv::NickChange(self.get_nick().to_owned()));
-        }
+        update_status!(
+            self,
+            status,
+            match status {
+                ConnStatus::Introduce { mut stream, .. } => {
+                    wire::user(&mut stream, &self.hostname, &self.realname).unwrap();
+                    wire::nick(&mut stream, &self.nicks[self.current_nick_idx]).unwrap();
+                    evs.push(ConnEv::NickChange(self.get_nick().to_owned()));
+                    ConnStatus::PingPong {
+                        ticks_passed: 0,
+                        stream: stream,
+                    }
+                }
+                _ =>
+                    status,
+            }
+        );
 
         if let Msg {
             cmd: Cmd::JOIN { .. },
@@ -642,16 +643,16 @@ impl<'poll> Conn<'poll> {
             ..
         } = msg
         {
-            // RPL_ENDOFMOTD. Join auto-join channels.
-            for chan in &self.auto_join {
-                wire::join(&mut self.out_buf, chan).unwrap();
-                reregister_for_rw(self.poll, self.get_raw_fd());
-            }
+            if let Some(mut stream) = self.status.get_stream_mut() {
+                // RPL_ENDOFMOTD. Join auto-join channels.
+                for chan in &self.auto_join {
+                    wire::join(&mut stream, chan).unwrap();
+                }
 
-            // Set away mode
-            if let &Some(ref reason) = &self.away_status {
-                wire::away(&mut self.out_buf, Some(reason)).unwrap();
-                self.reregister_for_rw();
+                // Set away mode
+                if let Some(ref reason) = self.away_status {
+                    wire::away(stream, Some(reason)).unwrap();
+                }
             }
         }